+++ packages/integration-sdk-runtime/src/execution/jobState.ts
@@ -142,6 +142,31 @@ export function createStepJobState({
       });
     },
 
+    findEntity: async (_key: string) => {
+      const graphObjectMetadata = duplicateKeyTracker.getGraphObjectMetadata(
+        _key,
+      );
+
+      if (!graphObjectMetadata) {
+        return null;
+      }
+
+      try {
+        const entity = await graphObjectStore.getEntity({
+          _key,
+          _type: graphObjectMetadata._type,
+        });
+
+        return entity;
+      } catch (err) {
+        if (err instanceof IntegrationMissingKeyError) {
+          return null;
+        }
+
+        throw err;
+      }
+    },
+
     iterateEntities: (filter, iteratee) =>
       graphObjectStore.iterateEntities(filter, iteratee),